requirements.txt
matplotlib==3.5.1
numpy==1.21.5
scipy==1.7.3
tensorflow_gpu==1.14.0
